+++ lib/cli/ConsoleInteraction.ts
@@ -1,6 +1,6 @@
 import Board from '../Board';
-import { IO } from '../IO';
-import { Display } from '../Display';
+import { IO } from '../interfaces/IO';
+import { Display } from '../interfaces/Display';
 import 'core-js/stable';
 import 'regenerator-runtime/runtime';
