+++ lib/interfaces/Player.ts
@@ -1,4 +1,4 @@
-import Board from './Board';
+import Board from '../Board';
 
 export interface Player {
   getMove(board: Board): Promise<number>;
